fixed duplicate issues for favorites

+++ assets/js/recipe.js
@@ -5,7 +5,7 @@ let ingredientsArray = JSON.parse(localStorage.getItem('ingredients')) || []
 let savedRecipe = JSON.parse(localStorage.getItem('savedRecipe')) || []
 let localArray = JSON.parse(localStorage.getItem('savedRecipe')) || []
 console.log(localArray)
-let allRecipes=[]
+let allRecipes = []
 
 //array of ingredients user inputed
 
@@ -77,12 +77,12 @@ function renderSlides(ingredients) {
 
             //setting data = an array within res
             let data = res.data
-            
+
             //creating a card
             let card = document.createElement('div')
             // console.log(data)
             //looping through 5 reesults, creating cards for each and then getting more information from spoonacular based of of the recipes id, previouse search reults do not hold all of the infromation required.
-            let saved =0
+            let saved = 0
             for (i = 0; i < data.length; i++) {
                 const card = document.createElement('div')
 
@@ -123,15 +123,16 @@ function renderSlides(ingredients) {
               <div>${recipe.instructions}</div>
               <button class="uk-modal-close uk-button-default uk-button-large" type="button">Close</button>
              <button class="uk-button uk-button-primary uk-button-large save${saved++}" type="button">Save</button>
+            
               </div>
 
               </div>
       
                 `
-                        
+
 
                         document.getElementById('slider').append(listItem)
-                       
+
 
 
 
@@ -154,24 +155,52 @@ document.addEventListener('click', event => {
 
     for (i = 0; i < allRecipes.length; i++) {
         if (event.target.classList.contains(`save${i}`)) {
-         
-        console.log(allRecipes[i])
+            if (event.target.innerHTML === 'Save') {
 
-          
-            let recipeObject = {}
-            recipeObject["name"] = allRecipes[i].title
-            recipeObject["image"] = allRecipes[i].image
-            recipeObject["id"] = allRecipes[i].id
-            recipeObject["summary"] = allRecipes[i].summary
-            recipeObject["instructions"] = allRecipes[i].instructions
 
-            console.log(recipeObject)
-            localArray.push(recipeObject);
+               event.target.innerHTML= 'Saved'
+                console.log(allRecipes[i])
 
-            console.log(localArray)
-          
-    
-            localStorage.setItem('savedRecipe', JSON.stringify((localArray)))
+
+                let recipeObject = {}
+                recipeObject["name"] = allRecipes[i].title
+                recipeObject["image"] = allRecipes[i].image
+                recipeObject["id"] = allRecipes[i].id
+                recipeObject["summary"] = allRecipes[i].summary
+                recipeObject["instructions"] = allRecipes[i].instructions
+                let alreadySaved = false
+                localArray.forEach(element =>{
+                    if(element ===recipeObject)
+                    {
+                        console.log('equal')
+                        alreadySaved = true
+                    }
+
+
+                })
+
+
+
+                if(alreadySaved=== false){
+                    console.log(recipeObject)
+                    localArray.push(recipeObject);
+
+                    console.log(localArray)
+
+
+                    localStorage.setItem('savedRecipe', JSON.stringify((localArray)))
+                }
+
+            
+
+
+            }
+            else if(event.target.innerHTML==='Saved'){
+                console.log('already saved')
+            }
+
+
+
         }
     }
